Conflicts:
	perl/openprint/Paper.pm
	perl/openprint/ScheduledJob.pm
	perl/openprint/Skid.pm
	www/public_html/administrator/managerial/company_profiles.html
	www/public_html/administrator/reports/projects.html
	www/public_html/employee/accounting/_stock.html
	www/public_html/employee/accounting/stock.html
	www/public_html/employee/inventory/manifests.html
	www/public_html/skins/PointOne Graphics Inc/css/common.css
	www/public_html/skins/PointOne Graphics Inc/layouts/employee/default.html

+++ perl/openprint/employee_purchase_order.pm
@@ -475,8 +475,12 @@ sub history {
 		my $PO = new openprint::PurchaseOrder( $param{'po_id'} );
 		$variable{'error'} .= $PO->send_to_vendor();
 		delete $param{'po_id'};
+	} elsif ( $param{'btnFunction'} eq 'Email Me' ) {
+		my $PO = new openprint::PurchaseOrder( $param{'po_id'} );
+		$variable{'error'} = $PO->send_to_me();
+		delete $param{'po_id'};
 	} # end if
-	ssi::save_params( '/employee/purchase_order/history.html', ( 'starting_start_year','starting_start_month','starting_start_day','starting_end_year','starting_end_month','starting_end_day','authorized', 'supplier_id','created_by','deleted','types', 'item_id', 'cancelled' ) );
+	_history();
 	ssi::setup_date_select( '/employee/purchase_order/history.html', 'starting_start', -7 );
 	ssi::setup_date_select( '/employee/purchase_order/history.html', 'starting_end', '' );
 	$session{'/employee/purchase_order/history.html?cancelled'} = '0' if ! exists $session{'/employee/purchase_order/history.html?cancelled'};
@@ -520,9 +524,12 @@ sub _po_content_line {
 		$variable{'error'} .= $PO->save();
 	} elsif ( $param{'action'} eq 'delete' ) {
 		my $PO_Content = new openprint::PurchaseOrder_Content( $param{'id'} );
-		$PO = $PO_Content->PurchaseOrder();
-		$PO_Content->delete();
-		$variable{'error'} .= $PO->save();
+		if ( $PO_Content->id() ) {
+			# Might have already been deleted
+			$PO = $PO_Content->PurchaseOrder();
+			$PO_Content->delete();
+			$variable{'error'} .= $PO->save();
+		} # end if
 	} # end if
 } # end sub _purchase_order_content_line
